requirements.txt
pytest-playwright
pytest-parallel-39
pytest-xdist